zeta2 cli: Split retrieval stats module
